+++ src/wai/annotations/format/void/specifier/__init__.py
@@ -1,3 +1,4 @@
+from ._VoidACOutputFormatSpecifier import VoidACOutputFormatSpecifier
 from ._VoidICOutputFormatSpecifier import VoidICOutputFormatSpecifier
 from ._VoidISOutputFormatSpecifier import VoidISOutputFormatSpecifier
 from ._VoidODOutputFormatSpecifier import VoidODOutputFormatSpecifier