+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 - Added new audio domain for classification using suffix `-ac`
 - Added dataset reader for audio files: `from-audio-files-sp`, `from-audio-files-ac`
 - Added dataset writer for audio files: `to-audio-files-sp`, `to-audio-files-ac`
+- Added dummy sink for audio files: `to-void-ac`
 
 
 0.1.7 (2022-06-13)